Consolidate extrema postaggregation files using Go generics - Eliminate code duplication for double/long greatest/least operations - Use Go generics with type aliases for backward compatibility - Maintain 100% API compatibility and test coverage
